+++ src/users/models/__init__.py
@@ -1,0 +1,5 @@
+from .custom_user import CustomUser
+from .client import Client
+from .admin import Admin
+from credit_card import CreditCard
+from address import Address
