+++ src/users/models/custom_user.py
@@ -1,0 +1,57 @@
+from django.db import models
+from django.contrib.auth.models import (
+    AbstractUser,
+    BaseUserManager,
+    PermissionsMixin,
+)
+
+# from .validators import validate_card_number
+import uuid
+
+
+class CustomUserManager(BaseUserManager):
+    """ """
+
+    def create_user(self, email, password, **extra_fields):
+        if not email:
+            raise ValueError("The Email field must be set")
+        email = self.normalize_email(email)
+        user = self.model(email=email, **extra_fields)
+        user.set_password(password)
+        user.save(using=self._db)
+        return user
+
+    def create_superuser(self, email, password, **extra_fields):
+        extra_fields.setdefault("is_superuser", True)
+        extra_fields.setdefault("is_staff", True)
+
+        if extra_fields.get("is_superuser") is not True:
+            raise ValueError("Superuser must have is_superuser=True.")
+        if extra_fields.get("is_staff") is not True:
+            raise ValueError("Superuser must have is_staff=True.")
+
+        return self.create_user(email, password, **extra_fields)
+
+
+class CustomUser(AbstractUser, PermissionsMixin):
+    """ """
+
+    GENDER_CHOICES = (
+        ("M", "Male"),
+        ("F", "Female"),
+        ("O", "Other"),
+        ("N", "Prefer not to say"),
+    )
+
+    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
+    first_name = models.CharField(max_length=30, blank=True)
+    last_name = models.CharField(max_length=30, blank=True)
+    gender = models.CharField(max_length=1, choices=GENDER_CHOICES, default="O")
+    email = models.EmailField(unique=True)
+    username = models.CharField(max_length=150, unique=True, null=True)
+    is_active = models.BooleanField(default=True)
+    is_staff = models.BooleanField(default=False)
+    is_superuser = models.BooleanField(default=False)
+
+    class Meta:
+        abstract = True
